Add user-specific filtering for API keys and typeorm dependency
This commit introduces a dependency on the typeorm package and modifies methods related to API key manipulation. Now a user-specific filtering is implemented when getting API keys, meaning a user will only see the keys that were created by them. This change is necessary for security reasons to prevent users from accessing other users' keys. In addition, userid property has been added when creating new API keys. This is needed for the new user-specific filtering functionality to work. The typeorm dependency has been added for future integration with a database.