Add premium subscription feature with Firebase
This update incorporated firebase functions to track the premium subscription status of users. Redux-thunk was imported for asynchronous actions and used in fetching premium status from Firebase. A new reducer - checkoutReducer was added to handle the state of the checkout modal. Modifications were also made in the App.js to only allow checkout modal when the user is authenticated. Firebase analytics was integrated to track when checkout was clicked and a purchase was made. Important changes also include a fallback to direct users to the checkout modal when they try to access premium features without a subscription.