Add Facebook login option to OAuth
This commit adds an option for users to sign in with their Facebook account. Specifically, it imports the FacebookAuthProvider from 'firebase/auth' in our LoginForm view. Additionally, it introduces a FacebookIcon from '@mui/icons-material/Facebook' and a associated StyledButton for triggering the OAuth sign-in process with Facebook. Adding this feature provides users with another convenient way to securely log in to our platform. This implementation increases the number of potential users who prefer using their Facebook account for a quick login process.